fix integer overflow with many scans.

+++ experiments/FID/__init__.py
@@ -123,7 +123,7 @@ class Experiment(BaseExperiment): # must be named 'Experiment'
 
     def raw_data(self):
         data = self.programs['FID'].data
-        data = data.astype(np.float32).view(np.complex64)
+        data = data.view(np.complex64)
         return data
 
     def autophase(self, data):
