+++ experiments/Noise/__init__.py
@@ -89,4 +89,4 @@ class Experiment(BaseExperiment): # must be named 'Experiment'
 
     def raw_data(self):
         data = self.programs['Noise'].data
-        return data.astype(np.float32).view(np.complex64)
+        return data.view(np.complex64)
